+++ internal/connectors/easyship/lib.go
@@ -12,23 +12,22 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package jsonconverter
+package easyship
 
-import (
-	sdk "github.com/wakflo/go-sdk/connector"
-)
+import sdk "github.com/wakflo/go-sdk/connector"
 
 func NewConnector() (*sdk.ConnectorPlugin, error) {
 	return sdk.CreateConnector(&sdk.CreateConnectorArgs{
-		Name:        "JSON",
-		Description: "Convert JSON to text and vice versa",
-		Logo:        "mdi:code-json",
+		Name:        "EasyShip",
+		Description: "shipping platform that connects sellers and marketplaces to couriers.",
+		Logo:        "fxemoji:diamondshapewithdotinside",
 		Version:     "0.0.1",
-		Category:    sdk.Tools,
+		Category:    sdk.Apps,
 		Authors:     []string{"Wakflo <integrations@wakflo.com>"},
 		Triggers:    []sdk.ITrigger{},
 		Operations: []sdk.IOperation{
-			NewConvertToJSONOperation(),
+			NewCreateNewLabelOperation(),
+			NewCreateCourierPickupOperation(),
 		},
 	})
 }